+++ app/handlers/tests/test_handlers_connector.py
@@ -1,4 +1,3 @@
-import collections
 from unittest.mock import ANY, MagicMock, patch
 
 import pytest
@@ -16,57 +15,6 @@ def mock_connector_spec_get_image():
     with patch("app.crds.ConnectorSpec.get_image") as mock_get_image:
         mock_get_image.return_value = "twingate/connector:test"
         yield mock_get_image
-
-
-@pytest.fixture()
-def k8s_client_mock():
-    client_mock = MagicMock()
-    with patch("kubernetes.client.CoreV1Api") as k8sclient_mock:
-        k8sclient_mock.return_value = client_mock
-        yield client_mock
-
-
-@pytest.fixture()
-def kopf_info_mock():
-    with patch("kopf.info") as kopf_info_mock:
-        yield kopf_info_mock
-
-
-HandlerRunnerResult = collections.namedtuple(
-    "HandlerRunnerResult",
-    [
-        "result",
-        "memo_mock",
-        "logger_mock",
-        "patch_mock",
-        "k8s_client_mock",
-        "kopf_info_mock",
-    ],
-)
-
-
-@pytest.fixture()
-def kopf_handler_runner(k8s_client_mock, kopf_info_mock):
-    def run(handler_f, crd, memo_mock, namespace="default"):
-        logger_mock = MagicMock()
-
-        patch_mock = MagicMock()
-        patch_mock.spec = {}
-        patch_mock.meta = {}
-
-        result = handler_f(
-            body=crd.model_dump(by_alias=True),
-            spec=crd.spec.model_dump(by_alias=True),
-            memo=memo_mock,
-            logger=logger_mock,
-            namespace=namespace,
-            patch=patch_mock,
-        )
-        return HandlerRunnerResult(
-            result, memo_mock, logger_mock, patch_mock, k8s_client_mock, kopf_info_mock
-        )
-
-    return run
 
 
 @pytest.fixture()
